+++ src/simmate/workflow_engine/README.md
@@ -3,4 +3,4 @@
 
 This module defines common workflow tasks, error handling, and job restarts. It is ment for users that want to create custom workflows from scratch. Use of this module is closely tied with [Tutorials 08 and 09](https://github.com/jacksund/simmate/tree/main/tutorials).
 
-This module is only meant for advanced users. Beginners should instead start by checking if there is already a workflow built for them (in `simmate.workflows`) or by checking if there are common tasks already built for the program they are using in `simmate.calculators` (for example, VASP users can check `simmate.calculators.vasp`).
+This module is only meant for advanced users. Beginners should instead start by checking if there is already a workflow built for them (in `simmate.workflows`) or by checking if there are common tasks already built for the program they are using in `simmate.calculators`. For example, VASP users can check `simmate.calculators.vasp`.
