+++ src/simmate/workflow_engine/__init__.py
@@ -4,8 +4,8 @@ from simmate.utilities import get_doc_from_readme
 
 __doc__ = get_doc_from_readme(__file__)
 
-from .tasks.supervised_staged_shell_task import S3Task
-from .tasks.workflow_task import WorkflowTask
+from .supervised_staged_shell_task import S3Task
+from .workflow_task import WorkflowTask
 from .workflow import Workflow, Parameter, ModuleStorage
 from .error_handler import ErrorHandler
